+++ src/main/java/com/gistpetition/api/petition/domain/repository/PetitionRepository.java
@@ -1,5 +1,6 @@
-package com.gistpetition.api.petition.domain;
+package com.gistpetition.api.petition.domain.repository;
 
+import com.gistpetition.api.petition.domain.Petition;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
