+++ src/test/java/com/gistpetition/api/acceptance/petition/CreatePetitionAcceptanceTest.java
@@ -3,7 +3,7 @@ package com.gistpetition.api.acceptance.petition;
 import com.gistpetition.api.acceptance.AcceptanceTest;
 import com.gistpetition.api.acceptance.common.TUser;
 import com.gistpetition.api.petition.domain.Category;
-import com.gistpetition.api.petition.domain.PetitionRepository;
+import com.gistpetition.api.petition.domain.repository.PetitionRepository;
 import com.gistpetition.api.petition.dto.PetitionRequest;
 import com.gistpetition.api.user.domain.UserRole;
 import io.restassured.response.Response;
